Extract stateful EventTile adapters
